API/requirements.txt
numpy==1.26.4
fastapi[standard]
uvicorn[standard]
python-dotenv
llama-index-core
llama-index-vector-stores-weaviate
llama-index-llms-openai
langchain-text-splitters
weaviate-client
# sentence-transformers
langchain
langchain-openai
colorama
httpx
cohere